+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.132] - 2026-05-26
+- [Maintenance Calendar]
+  - New Module [Iam User Access Key Age]
+
 ## [1.0.131] - 2026-05-15
 - [Proto Instance]
   - Fix management Lambda failing to unmount /home when users are logged in